MVC 적용
